Add finding intersection of two array problem.
Add solution for finding intersection of two arrays.
One is a brute force solution that runs at O(M*N).
Another is a two pointer method that uses sorting of arrays and that
runs in O(NlogN) time as sorting is the worst case run time.
Two pointers are used to traverse each list and adds a unique element to
the resulting list.